CanLog_Analisis/requirements.txt
PySide6>=6.5.0
python-can>=4.3.0
canmatrix>=1.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
